+++ api/CHANGELOG.md
@@ -52,6 +52,7 @@
 - Made `Any package` tags not searchable for the `View patches` screens patch widgets (and also the `Any version` tag when the `Any package` tag exists with it)
 - Added an update notice tag to saved patched apps when the imported patch bundle version is newer than the one used to patch the app https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/163
 - Updated Patch Bundle Discovery to use the new `api/v2` & `latest?channel=` URLs while keeping backwards compatibility with `api/v1` URLs
+- Added a draggable transparency adjustment bar to Settings > General for when a image is set as the background
 
 
 # Bug fixes